Keyboard driver upgrade
Added support for extended scan codes. This will get the arrow keys when they are pressed. They are currently outputting weird other keys but work in progress.

Fixed a bug in ksh where the first line would get assigned the address of the "\0" string.

Added a write outputting command. Still work in progress.

+++ keyboard.h
@@ -151,7 +151,7 @@ typedef uint16_t keycode;
 
 struct keyboard_state
 {
-    uint8_t currentScancode;
+    uint16_t currentScancode;
     uint16_t currentKeycode;
     
     uint8_t* keyStates;
